Fix : FriendFragment 1차 수정본
오류 : EditText 부분에서 focus changed될 때 키보드 부분이 내려가지 않음 -> 다시 해보고 안되면 다른 view로 고쳐서 다시 실행해봄
