+++ android/app/src/main/java/com/example/jonggangtime/UI/Friends/ListFriendsFragment.kt
@@ -1,27 +1,16 @@
-/*
 package com.example.jonggangtime.UI.Friends
 
-import android.content.Context
-import android.graphics.Rect
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
-import android.view.ViewGroup
-import android.view.inputmethod.InputMethodManager
-import androidx.core.content.ContextCompat.getSystemService
 import com.example.jonggangtime.R
 import com.example.jonggangtime.Utils.BaseFragment
-import com.example.jonggangtime.databinding.FragmentDetailFriendBinding
 import com.example.jonggangtime.databinding.FragmentListFriendsBinding
-import com.google.android.material.textfield.TextInputEditText
 
 
 class ListFriendsFragment : BaseFragment<FragmentListFriendsBinding>(FragmentListFriendsBinding::inflate) {
 
     override fun initAfterBinding() {
-        TODO("Not yet implemented")
+
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -29,13 +18,21 @@ class ListFriendsFragment : BaseFragment<FragmentListFriendsBinding>(FragmentLis
 
         initNavigation()
 
-        binding.searchTf.setOnFocusChangeListener { v, hasFocus ->
+        binding.friendsSearchEt.setOnFocusChangeListener { v, hasFocus ->
             if(hasFocus){
                 childFragmentManager.beginTransaction()
                     .replace(R.id.friend_list_fl, SeekListFriendsFragment())
                     .commitAllowingStateLoss()
             }
         }
+
+        /*binding.searchTf.setOnFocusChangeListener { v, hasFocus ->
+            if(hasFocus){
+                childFragmentManager.beginTransaction()
+                    .replace(R.id.friend_list_fl, SeekListFriendsFragment())
+                    .commitAllowingStateLoss()
+            }
+        }*/
     }
 
     private fun initNavigation() {
@@ -43,4 +40,4 @@ class ListFriendsFragment : BaseFragment<FragmentListFriendsBinding>(FragmentLis
             .replace(R.id.friend_list_fl, AlreadyListFiendsFragment())
             .commitAllowingStateLoss()
     }
-}*/
+}
