+++ android/app/src/main/res/layout/fragment_already_list_fiends.xml
@@ -1,5 +1,142 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView
+<androidx.core.widget.NestedScrollView android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/friends_main_layout"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <!--친구 요청 화면-->
+            <com.google.android.material.card.MaterialCardView
+                android:id="@+id/friends_request_layout"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:strokeColor="@color/gray"
+                android:backgroundTint="@color/red_0"
+                app:cardCornerRadius="5dp"
+                app:strokeWidth="1dp"
+                app:cardElevation="0dp"
+                android:layout_marginHorizontal="15dp"
+                android:layout_marginBottom="20dp"
+                android:layout_marginTop="15dp">
+
+                <androidx.constraintlayout.widget.ConstraintLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent">
+
+                    <TextView
+                        android:id="@+id/friends_recieved_title_tv"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="수락 대기 중인 요청"
+                        android:textColor="@color/black"
+                        android:textStyle="bold"
+                        android:textSize="16sp"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        android:layout_marginStart="20dp"
+                        android:layout_marginTop="8dp"/>
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/friends_recieved_rv"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_marginHorizontal="20dp"
+                        app:layout_constraintTop_toBottomOf="@id/friends_recieved_title_tv"
+                        android:layout_marginTop="10dp"
+                        android:overScrollMode="never"
+                        android:nestedScrollingEnabled="false"
+                        />
+
+                    <TextView
+                        android:id="@+id/friends_send_title_tv"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="내가 보낸 요청"
+                        android:textColor="@color/black"
+                        android:textStyle="bold"
+                        android:textSize="16sp"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toBottomOf="@id/friends_recieved_rv"
+                        android:layout_marginStart="20dp"
+                        android:layout_marginTop="13dp"/>
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/friends_send_rv"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_marginHorizontal="20dp"
+                        app:layout_constraintTop_toBottomOf="@id/friends_send_title_tv"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        android:layout_marginTop="10dp"
+                        android:layout_marginBottom="10dp"
+                        android:overScrollMode="never"
+                        android:nestedScrollingEnabled="false"
+                        />
+
+                </androidx.constraintlayout.widget.ConstraintLayout>
+
+            </com.google.android.material.card.MaterialCardView>
+
+            <!--내 친구-->
+            <com.google.android.material.card.MaterialCardView
+                android:id="@+id/friends_my_friend_layout"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                app:layout_constraintTop_toBottomOf="@id/friends_request_layout"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:strokeColor="@color/gray"
+                app:cardCornerRadius="5dp"
+                app:strokeWidth="1dp"
+                app:cardElevation="0dp"
+                android:layout_marginHorizontal="15dp"
+                android:layout_marginBottom="20dp"
+                android:layout_marginTop="20dp">
+
+                <androidx.constraintlayout.widget.ConstraintLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent">
+
+                    <TextView
+                        android:id="@+id/friends_my_friend_title"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="내 친구"
+                        android:textColor="@color/black"
+                        android:textStyle="bold"
+                        android:textSize="16sp"
+                        app:layout_constraintStart_toStartOf="parent"
+                        app:layout_constraintTop_toTopOf="parent"
+                        android:layout_marginStart="20dp"
+                        android:layout_marginTop="8dp"/>
+
+                    <androidx.recyclerview.widget.RecyclerView
+                        android:id="@+id/friends_my_friend_rv"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_marginHorizontal="20dp"
+                        app:layout_constraintTop_toBottomOf="@id/friends_my_friend_title"
+                        app:layout_constraintBottom_toBottomOf="parent"
+                        android:layout_marginTop="10dp"
+                        android:layout_marginBottom="10dp"
+                        android:overScrollMode="never"
+                        android:nestedScrollingEnabled="false"
+                        />
+
+                </androidx.constraintlayout.widget.ConstraintLayout>
+
+            </com.google.android.material.card.MaterialCardView>
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.core.widget.NestedScrollView>
+<!--<ScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -81,4 +218,4 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
-</ScrollView>
+</ScrollView>-->
