+++ android/app/src/main/res/layout/item_friend.xml
@@ -18,38 +18,4 @@
         android:textSize="16sp"
         android:textColor="@color/black"/>
 
-    <!--<TextView
-        android:id="@+id/friend_nickname_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:text="닉네임"/>
-    <TextView
-        android:id="@+id/friend_name_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toEndOf="@id/friend_nickname_tv"
-        app:layout_constraintBottom_toBottomOf="@id/friend_nickname_tv"
-        android:text="이름"/>
-
-    <ImageView
-        android:id="@+id/check_iv"
-        android:layout_width="18dp"
-        android:layout_height="18dp"
-        app:layout_constraintEnd_toStartOf="@id/cancel_iv"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:src="@drawable/icon_blue_check"/>
-
-    <ImageView
-        android:id="@+id/cancel_iv"
-        android:layout_width="18dp"
-        android:layout_height="18dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:src="@drawable/icon_red_close"/>-->
-
 </androidx.constraintlayout.widget.ConstraintLayout>
